requirements.txt
playwright==1.40.0
pandas>=2.2.0
python-dotenv==1.0.0
cryptography==42.0.5 